@@ -36,37 +36,36 @@ package edu.iu.uits.lms.canvasnotifier.model;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 import com.fasterxml.jackson.databind.annotation.JsonAppend;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import jakarta.persistence.Lob;
+import jakarta.persistence.NamedQueries;
+import jakarta.persistence.NamedQuery;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.SequenceGenerator;
+import jakarta.persistence.Table;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
 import lombok.ToString;
-import org.hibernate.annotations.Type;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import javax.persistence.Lob;
-import javax.persistence.NamedQueries;
-import javax.persistence.NamedQuery;
-import javax.persistence.OneToMany;
-import javax.persistence.SequenceGenerator;
-import javax.persistence.Table;
 import java.util.List;
 
 @Entity
 @Table(name = "CANVASNOTIFIER_JOB")
-@NamedQueries({
-        @NamedQuery(name = "Job.findAllRunningJobs", query = "from Job where status in ('STARTED', 'RESTARTED') order by id"),
-
-// 300 = 5 minutes in below query
-        @NamedQuery(name = "Job.getElevatedJobsOlderThan", query = "from Job job where job.senderWasElevated = true and job.senderIsElevated = true and sysdate - 300/(24*60*60) > job.modifiedOn order by job.id asc"),
-        @NamedQuery(name = "Job.getRunningJobsBySenderCanvasId", query = "from Job job where job.sender_canvasid = :senderId and ((status = 'PENDING' and sysdate - 300/(24*60*60) < job.modifiedOn) or status = 'STARTED' or status = 'RESTARTED') order by job.id asc")
-})
+//@NamedQueries({
+//        @NamedQuery(name = "Job.findAllRunningJobs", query = "from Job where status in ('STARTED', 'RESTARTED') order by id"),
+//
+//// 300 = 5 minutes in below query
+//        @NamedQuery(name = "Job.getElevatedJobsOlderThan", query = "from Job job where job.senderWasElevated = true and job.senderIsElevated = true and EXTRACT(EPOCH FROM CURRENT_TIMESTAMP) - 300 > EXTRACT(EPOCH FROM job.modifiedOn) order by job.id asc"),
+//        @NamedQuery(name = "Job.getRunningJobsBySenderCanvasId", query = "from Job job where job.sender_canvasid = :senderId and ((status = 'PENDING' and EXTRACT(EPOCH FROM CURRENT_TIMESTAMP) - 300 < EXTRACT(EPOCH FROM job.modifiedOn)) or status = 'STARTED' or status = 'RESTARTED') order by job.id asc")
+//})
 
 @SequenceGenerator(name = "CANVASNOTIFIER_JOB_ID_SEQ", sequenceName = "CANVASNOTIFIER_JOB_ID_SEQ", allocationSize = 1)
 @Data
@@ -92,12 +91,12 @@ public class Job extends ModelWithDates {
 
     @NonNull
     @Lob
-    @Type(type="text")
+    @Column(columnDefinition = "text")
     private String body;
 
     @NonNull
     @Lob
-    @Type(type="text")
+    @Column(columnDefinition = "text")
     private String json_csv;
 
     @NonNull
