@@ -34,6 +34,8 @@ package edu.iu.uits.lms.canvasnotifier.repository;
  */
 
 import edu.iu.uits.lms.canvasnotifier.model.Job;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.ListCrudRepository;
 import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Component;
@@ -41,8 +43,17 @@ import org.springframework.stereotype.Component;
 import java.util.List;
 
 @Component
-public interface JobRepository extends PagingAndSortingRepository<Job, Long> {
+public interface JobRepository extends PagingAndSortingRepository<Job, Long>, ListCrudRepository<Job, Long> {
+
+    @Query("from Job where status in ('STARTED', 'RESTARTED') order by id")
     List<Job> findAllRunningJobs();
+
+    // 300 = 5 minutes in below query
+//    @Query("from Job job where job.senderWasElevated = true and job.senderIsElevated = true and EXTRACT(EPOCH FROM CURRENT_TIMESTAMP) - 300 > EXTRACT(EPOCH FROM job.modifiedOn) order by job.id asc")
+    @Query(nativeQuery = true, value = "from Job job where job.senderWasElevated = true and job.senderIsElevated = true and CURRENT_TIMESTAMP - INTERVAL '5' MINUTE > job.modifiedOn order by job.id asc")
     List<Job> getElevatedJobsOlderThan();
+
+//    @Query("from Job job where job.sender_canvasid = :senderId and ((status = 'PENDING' and EXTRACT(EPOCH FROM CURRENT_TIMESTAMP) - 300 < EXTRACT(EPOCH FROM job.modifiedOn)) or status = 'STARTED' or status = 'RESTARTED') order by job.id asc")
+    @Query(nativeQuery = true, value = "from Job job where job.sender_canvasid = :senderId and ((status = 'PENDING' and CURRENT_TIMESTAMP - INTERVAL '5' MINUTE < job.modifiedOn) or status = 'STARTED' or status = 'RESTARTED') order by job.id asc")
     List<Job> getRunningJobsBySenderCanvasId(@Param("senderId") String senderId);
 }
