@@ -1,61 +1,61 @@
-package edu.iu.uits.lms.canvasnotifier.services.swagger;
-
-/*-
- * #%L
- * canvasnotifier
- * %%
- * Copyright (C) 2015 - 2022 Indiana University
- * %%
- * Redistribution and use in source and binary forms, with or without modification,
- * are permitted provided that the following conditions are met:
- *
- * 1. Redistributions of source code must retain the above copyright notice, this
- *    list of conditions and the following disclaimer.
- *
- * 2. Redistributions in binary form must reproduce the above copyright notice,
- *    this list of conditions and the following disclaimer in the documentation
- *    and/or other materials provided with the distribution.
- *
- * 3. Neither the name of the Indiana University nor the names of its contributors
- *    may be used to endorse or promote products derived from this software without
- *    specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
- * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED.
- * IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
- * INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
- * BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
- * LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE
- * OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED
- * OF THE POSSIBILITY OF SUCH DAMAGE.
- * #L%
- */
-
-import edu.iu.uits.lms.canvasnotifier.amqp.CanvasNotifierMessageListener;
-import edu.iu.uits.lms.iuonly.services.CanvasDataServiceImpl;
-import edu.iu.uits.lms.iuonly.services.SisServiceImpl;
-import org.springframework.boot.context.metrics.buffering.BufferingApplicationStartup;
-import org.springframework.boot.test.context.TestConfiguration;
-import org.springframework.boot.test.mock.mockito.MockBean;
-
-import javax.sql.DataSource;
-
-@TestConfiguration
-public class SwaggerNotifierTestConfig {
-    @MockBean
-    private BufferingApplicationStartup bufferingApplicationStartup;
-
-    @MockBean
-    private CanvasDataServiceImpl canvasDataService;
-
-    @MockBean
-    private CanvasNotifierMessageListener canvasNotifierMessageListener;
-
-    @MockBean
-    private DataSource dataSource;
-
-    @MockBean
-    private SisServiceImpl sisService;
-}
+//package edu.iu.uits.lms.canvasnotifier.services.swagger;
+//
+///*-
+// * #%L
+// * canvasnotifier
+// * %%
+// * Copyright (C) 2015 - 2022 Indiana University
+// * %%
+// * Redistribution and use in source and binary forms, with or without modification,
+// * are permitted provided that the following conditions are met:
+// *
+// * 1. Redistributions of source code must retain the above copyright notice, this
+// *    list of conditions and the following disclaimer.
+// *
+// * 2. Redistributions in binary form must reproduce the above copyright notice,
+// *    this list of conditions and the following disclaimer in the documentation
+// *    and/or other materials provided with the distribution.
+// *
+// * 3. Neither the name of the Indiana University nor the names of its contributors
+// *    may be used to endorse or promote products derived from this software without
+// *    specific prior written permission.
+// *
+// * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+// * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+// * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED.
+// * IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
+// * INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
+// * BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+// * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
+// * LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE
+// * OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED
+// * OF THE POSSIBILITY OF SUCH DAMAGE.
+// * #L%
+// */
+//
+//import edu.iu.uits.lms.canvasnotifier.amqp.CanvasNotifierMessageListener;
+//import edu.iu.uits.lms.iuonly.services.CanvasDataServiceImpl;
+//import edu.iu.uits.lms.iuonly.services.SisServiceImpl;
+//import org.springframework.boot.context.metrics.buffering.BufferingApplicationStartup;
+//import org.springframework.boot.test.context.TestConfiguration;
+//import org.springframework.boot.test.mock.mockito.MockBean;
+//
+//import javax.sql.DataSource;
+//
+//@TestConfiguration
+//public class SwaggerNotifierTestConfig {
+//    @MockBean
+//    private BufferingApplicationStartup bufferingApplicationStartup;
+//
+//    @MockBean
+//    private CanvasDataServiceImpl canvasDataService;
+//
+//    @MockBean
+//    private CanvasNotifierMessageListener canvasNotifierMessageListener;
+//
+//    @MockBean
+//    private DataSource dataSource;
+//
+//    @MockBean
+//    private SisServiceImpl sisService;
+//}
